run as few benchmarks as possible using content-addressing for measurements
note: this commit doesn't fully work, there are some path handling issues with benchmark binaries. will resolve in a followup.

we store the actual measurements under a hash of (benchmark binary, runner, cgroup config), and then write "pointers" to those hashes for the actual metadata we want to store.

on the assumption that most commits won't cause most benchmark binaries to change, this should allow us to run all of the benchmarks with far less hardware.

also, in a display of excellent software engineering, i'm including other changes too:

* deleting unused modules like estimates and plan
* removes a bunch of code from run_plan to clean up api boundaries