Add cargo alias for formatting core crates, block CI on them being formatted.
We can't run rustfmt on criterion because we don't want to generate a large diff with upstream -- ideally the fork will go away soon.

We can't run rustfmt on the bench crates because we don't want to require rustfmt to run the proc macro until it's 1.0 and part of the stable distribution.